+++ server/index.ts
@@ -1,7 +1,6 @@
-import express from "express";
+import express, { Router } from "express";
 import mongoose from "mongoose";
 import "dotenv/config";
-import userApis from "./src/v1/api/userApi";
 
 // Express FWによるローカルサーバーの立ち上げ
 const app: express.Express = express();
@@ -11,6 +10,9 @@ const url = process.env.MONGODB_URL ? process.env.MONGODB_URL : "";
 // jsonオブジェクトを扱うため
 app.use(express.json());
 
+// エンドポイント
+app.use("/api/v1", require("./src/v1/routes/auth"));
+
 // DB接続
 try {
 	mongoose.set("strictQuery", true);
@@ -19,9 +21,6 @@ try {
 } catch (e) {
 	console.log(e);
 }
-
-// ユーザーのAPI呼び出し
-userApis(app);
 
 app.listen(PORT, () => {
 	console.log("ローカルサーバー起動中");
